refactor: move explorer page widgets into feature ui
