+++ src/features/Explorer/ui/ExplorerWorkspace.tsx
@@ -1,17 +1,12 @@
-import { useAtomValue, useSetAtom } from 'jotai';
+import { useAtomValue } from 'jotai';
 import { rightPanelOpenAtom, rightPanelTypeAtom } from '@/entities/AppView/model/atoms';
 import { TabContainer } from '@/features/File/OpenFiles/ui/TabContainer';
-import { ExplorerSidebar } from '../../ExplorerSidebar/ui/ExplorerSidebar';
-import { WorkspacePanel } from '../../WorkspacePanel/ui/WorkspacePanel';
+import { ExplorerSidebar } from './ExplorerSidebar';
+import { WorkspacePanel } from './WorkspacePanel';
 
 export function ExplorerWorkspace() {
   const rightPanelOpen = useAtomValue(rightPanelOpenAtom);
   const rightPanelType = useAtomValue(rightPanelTypeAtom);
-  const setRightPanelOpen = useSetAtom(rightPanelOpenAtom);
-
-  function handleCloseWorkspacePanel() {
-    setRightPanelOpen(false);
-  }
 
   return (
     <div className="flex h-full w-full overflow-hidden">
@@ -21,7 +16,7 @@ export function ExplorerWorkspace() {
         <TabContainer />
       </div>
 
-      {rightPanelOpen && rightPanelType === 'workspace' && <WorkspacePanel onClose={handleCloseWorkspacePanel} />}
+      {rightPanelOpen && rightPanelType === 'workspace' && <WorkspacePanel />}
     </div>
   );
 }
