+++ src/pages/explorer/PageExplorer.tsx
@@ -1,4 +1,4 @@
-import { ExplorerWorkspace } from './widgets/ExplorerWorkspace/ui/ExplorerWorkspace';
+import { ExplorerWorkspace } from '@/features/Explorer/ui/ExplorerWorkspace';
 
 export function PageExplorer() {
   return <ExplorerWorkspace />;
